(ft-action-func-tutorial): move project from root folder to a specific folder to solve jenkins build error [FIX]
