+++ csharp-advanced-topics/ActionAndFuncDelegateTutorial/ActionAndFuncDelegateTutorial/ActionAndFuncDelegateTutorial.csproj
@@ -5,8 +5,4 @@
     <TargetFramework>net5.0</TargetFramework>
   </PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="xunit" Version="2.4.1" />
-  </ItemGroup>
-
 </Project>
