+++ csharp-advanced-topics/ActionAndFuncDelegateTutorial/ActionAndFuncDelegateTutorialTest/ActionAndFuncDelegateTutorialTest.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="xunit" Version="2.4.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\ActionAndFuncDelegateTutorial\ActionAndFuncDelegateTutorial.csproj" />
+  </ItemGroup>
+
+</Project>
